requirements.txt
requests>=2.14.0
pycryptodome>=3.6.6
